+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - List workspace information when selection fails.
 - Add `list` option to `command` parameter for `state` method.
 - Add `refreshOnly` parameter to `plan` method.
+- Add `display` parameter to `output` method.
 
 ### 1.6.0
 **Helm**